+++ src/fed_rag/knowledge_stores/no_encode/mcp/__init__.py
@@ -1,4 +1,4 @@
-from .source import MCPStreamableHttpKnowledgeSource
+from .sources import MCPStreamableHttpKnowledgeSource
 from .store import MCPKnowledgeStore
 
 __all__ = ["MCPKnowledgeStore", "MCPStreamableHttpKnowledgeSource"]
