+++ src/fed_rag/knowledge_stores/no_encode/mcp/sources/__init__.py
@@ -1,0 +1,3 @@
+from .streamable_http import MCPStreamableHttpKnowledgeSource
+
+__all__ = ["MCPStreamableHttpKnowledgeSource"]
