+++ src/fed_rag/knowledge_stores/no_encode/mcp/sources/streamable_http.py
